Pulling new tables from ACS 2010-2014
New ACS 2010-2014 tables have been added to existing lists (table_list
and drop_bg_list), including new variables and race breakouts of
existing variables. Certain tables have been deleted from drop_list to
accommodate newly pulled tables.

One warning appears in the log for B23002. Estimates in this table are
used to calculate the employed population aged 25-64. As far as I know,
this is the only table we can use to ascertain this variable. It's not
in NIDC and appears to be suppressed by the Census.

Early on, we thought errors in the acs_sf_2010_14_dc program were caused
by the length of the tables lists. The text file is a back-up of a
program Leah wrote to solve the errors (we didn't end up using it).