Created vars for cost burden by age, income; employment commute
Created vars for renter and homeowner cost-burdened households by age of household and income level of householder.

Also created vars for workers 16 and over; whether or not they work in or outside of their state of residence.

And vars for workers 16 and over who don't work at home, by commute time.